XRBK macro: use u16 for request minor opcodes
TODO: clearly u16 values can't fit in the metabyte.  Need to see how minor opcodes are actually encoded, and whether they can automatically be placed at all.

+++ xrbk_macro/src/definition/expansion/trait.rs
@@ -37,7 +37,7 @@ impl Request {
 						#major_opcode
 					};
 
-					const MINOR_OPCODE: Option<u8> = {
+					const MINOR_OPCODE: Option<u16> = {
 						#minor_opcode
 					};
